+++ relay-general/src/protocol/contexts/mod.rs
@@ -24,6 +24,7 @@ mod otel;
 pub use otel::*;
 
 use crate::types::{Annotated, FromValue, Object, Value};
+use crate::user_agent::{ClientHints, RawUserAgentInfo};
 
 /// Operation type such as `db.statement` for database queries or `http` for external HTTP calls.
 /// Tries to follow OpenCensus/OpenTracing's span types.
@@ -84,6 +85,19 @@ impl Context {
             Context::Otel(_) => Some(OtelContext::default_key()),
             Context::Other(_) => None,
         }
+    }
+}
+
+/// Trait to get the Context both from the user agent string and also the new client hints.
+/// With an automatically derived function which tries to first get the context from client hints,
+/// if that fails it tries for the user agent string.
+pub trait FromUserAgentInfo: Sized {
+    fn from_client_hints(client_hints: &ClientHints) -> Option<Self>;
+    fn from_user_agent(user_agent: &str) -> Option<Self>;
+
+    fn from_hints_or_ua(raw_info: &RawUserAgentInfo) -> Option<Self> {
+        Self::from_client_hints(&raw_info.client_hints)
+            .or_else(|| raw_info.user_agent.and_then(Self::from_user_agent))
     }
 }