+++ relay-general/src/user_agent.rs
@@ -36,7 +36,6 @@ fn get_user_agent_from_headers(headers: &Headers) -> Option<&str> {
             }
         }
     }
-
     None
 }
 
@@ -78,4 +77,57 @@ pub fn parse_device(user_agent: &str) -> Device {
 /// Defaults to an empty operating system.
 pub fn parse_os(user_agent: &str) -> OS {
     UA_PARSER.parse_os(user_agent)
+}
+
+/// The data container, which has both the user agent string and the client hints.
+///
+/// Useful for the scenarios where you will use either information from client hints if it exists,
+/// and if not fall back to user agent string.
+#[derive(Default, Debug)]
+pub struct RawUserAgentInfo<'a> {
+    /// The "old style" of a single UA string.
+    pub user_agent: Option<&'a str>,
+    /// User-Agent client hints.
+    pub client_hints: ClientHints<'a>,
+}
+
+/// The client hint variable names mirror the name of the "SEC-CH" headers, see
+/// '<https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers#user_agent_client_hints>'
+#[derive(Default, Debug)]
+pub struct ClientHints<'a> {
+    /// The client's OS, e.g. macos, android...
+    pub sec_ch_ua_platform: Option<&'a str>,
+    /// The version number of the client's OS.
+    pub sec_ch_ua_platform_version: Option<&'a str>,
+    /// Name of the client's web browser and its version.
+    pub sec_ch_ua: Option<&'a str>,
+    /// Device model, e.g. samsung galaxy 3.
+    pub sec_ch_ua_model: Option<&'a str>,
+}
+
+impl<'a> RawUserAgentInfo<'a> {
+    pub fn new(headers: &'a Headers) -> Self {
+        let mut contexts = Self::default();
+
+        for item in headers.iter() {
+            if let Some((ref o_k, ref v)) = item.value() {
+                if let Some(k) = o_k.as_str() {
+                    match k.to_lowercase().as_str() {
+                        "user-agent" => contexts.user_agent = v.as_str(),
+
+                        "sec-ch-ua" => contexts.client_hints.sec_ch_ua = v.as_str(),
+                        "sec-ch-ua-model" => contexts.client_hints.sec_ch_ua_model = v.as_str(),
+                        "sec-ch-ua-platform" => {
+                            contexts.client_hints.sec_ch_ua_platform = v.as_str()
+                        }
+                        "sec-ch-ua-platform-version" => {
+                            contexts.client_hints.sec_ch_ua_platform_version = v.as_str()
+                        }
+                        _ => {}
+                    }
+                }
+            }
+        }
+        contexts
+    }
 }